Move ScrollEnumerator out of legacy_client directory
Move the `ScrollEnumerator` class because it is used by site search
queries. Its old location implied that it was only used by advanced
search, which will eventually be deprecated.
